src/classes.py
class Product:
    name: str
    description: str
    price: float
    quantity: int

    def __init__(self, name: str, description: str, price: float, quantity: int):
        self.name = name
        self.description = description
        self.price = price
        self.quantity = quantity


class Category:

    category_count = 0
    product_count = 0

    name: str
    description: str
    products: list[object]

    def __init__(self, name: str, description: str, products: list[object]):
        self.name = name
        self.description = description
        self.products = products
        Category.category_count += 1
        Category.product_count += len(products)
